backend_apis/requirements.txt
# Use Python 3.11, 3.12, or 3.13. Python 3.14 may fail when building pydantic-core.
fastapi==0.126.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
pydantic==2.11.0
httpx==0.27.2
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
fastapi-cors==0.0.6
python-dotenv==1.0.1
